Create a `writeFile` method on write-fs.js
Allow derived implementations of write-fs to have their own file
writing strategy.

At the moment we retained the currnt implementation, where are all primitive
sets on a prox are serialized & written in order. This has per-prox tracking
overhead & delay but it insures that latter writes on an object win, and if
multiple objects are having values set each can write independently.

This is seen as a middle-ground approach:
1. We could instead put all write ops into a global queue, & insure total ordering.
   This would mean each prox doesn't need it's own write queue, likely saving som
   amount of memory.
2. We could try to order write by the specific property only. If a user sets foo.a
   then foo.b, we could find a way to dispatch both of these while also insuring
   that if the user writes foo.a again, that the first foo.a write doesn't execute
   after the latter foo.a causing inconsistency.
3. We could let the writes fall as they may. :)