Start fs manager.
Serialize out fs operations, such that we can checkpoint tasks
and insure order of operations.

Switching this to some kind of a ring buffer is probably a good step,
because queue is going to be heavily worked otherwise.

Longer term I'd like to start to think about transactions, such that
work can be done safely. But this may not really make sense to do
only at the fs layer- ought not the prox itself be able to transact
too?