Get aggro working again under fork().
I'm also feeling like this shared prox instance thing was a mistake,
though. I like that I can reuse handler! But that was honestly
created because I couldn't figure out how to deal with pipeline names.
Which in turn I feel like I might be able to deal with, now.